restructure app in a major way.
